rename : 패키지 생성
필요한 패키지 생성
AI, AIRepository 클래스 각 패키지로 이동
