--- src/main/java/com/sparta/msa_exam/orderservicepractice/domain/ai/AIRepository.java
@@ -1,8 +1,0 @@
-package com.sparta.msa_exam.orderservicepractice.domain.ai;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.UUID;
-
-public interface AIRepository extends JpaRepository<AI, UUID> {
-}
